started work on load run and changed the values back to normal in exaustive tune

+++ scripts/alpaca_neural_net.py
@@ -5,7 +5,7 @@ from tensorflow.keras.callbacks import ModelCheckpoint, TensorBoard
 from tensorflow.keras.mixed_precision import experimental as mixed_precision
 from sklearn import preprocessing
 from time_functions import get_time_string
-from environment import (test_var, reports_directory, random_seed, error_file, back_test_days, 
+from environment import (test_var, reports_directory, model_saveload_directory, random_seed, error_file, back_test_days, 
 save_logs)
 from alpaca_nn_functions import (load_data, create_model, predict, accuracy_score, plot_graph, 
 get_accuracy, nn_report, return_real_predict)
@@ -35,7 +35,8 @@ def decision_neural_net(
     OPTIMIZER=defaults['OPTIMIZER'], 
     BATCH_SIZE=defaults['BATCH_SIZE'], 
     EPOCHS=defaults['EPOCHS'],
-    PATIENCE=defaults['PATIENCE']):
+    PATIENCE=defaults['PATIENCE'],
+    SAVELOAD=defaults['SAVELOAD']):
 #description of these parameters located inside environment.py
 
     start_time = time.time()
@@ -43,7 +44,8 @@ def decision_neural_net(
         ticker, N_STEPS=N_STEPS, LOOKUP_STEP=LOOKUP_STEP, TEST_SIZE=TEST_SIZE, 
         N_LAYERS=N_LAYERS, CELL=CELL, UNITS=UNITS, DROPOUT=DROPOUT, 
         BIDIRECTIONAL=BIDIRECTIONAL, LOSS=LOSS, OPTIMIZER=OPTIMIZER, 
-        BATCH_SIZE=BATCH_SIZE, EPOCHS=EPOCHS, PATIENCE=PATIENCE
+        BATCH_SIZE=BATCH_SIZE, EPOCHS=EPOCHS, PATIENCE=PATIENCE,
+        SAVELOAD=SAVELOAD
     )
 
     end_time = time.time()
@@ -66,7 +68,8 @@ def tuning_neural_net(ticker, end_date,
     OPTIMIZER=defaults['OPTIMIZER'], 
     BATCH_SIZE=defaults['BATCH_SIZE'], 
     EPOCHS=defaults['EPOCHS'],
-    PATIENCE=defaults['PATIENCE']):
+    PATIENCE=defaults['PATIENCE'],
+    SAVELOAD=defaults['SAVELOAD']):
 #description of these parameters located inside environment.py
     
     data, model, test_acc, valid_acc, train_acc, mae = make_neural_net(
@@ -74,10 +77,37 @@ def tuning_neural_net(ticker, end_date,
         N_STEPS=N_STEPS, LOOKUP_STEP=LOOKUP_STEP, TEST_SIZE=TEST_SIZE, 
         N_LAYERS=N_LAYERS, CELL=CELL, UNITS=UNITS, DROPOUT=DROPOUT, 
         BIDIRECTIONAL=BIDIRECTIONAL, LOSS=LOSS, OPTIMIZER=OPTIMIZER, 
-        BATCH_SIZE=BATCH_SIZE, EPOCHS=EPOCHS, PATIENCE=PATIENCE
+        BATCH_SIZE=BATCH_SIZE, EPOCHS=EPOCHS, PATIENCE=PATIENCE,
+        SAVELOAD=SAVELOAD
     )
     
     return test_acc, mae
+
+def saveload_neural_net(ticker, end_date, 
+    N_STEPS=defaults['N_STEPS'], 
+    LOOKUP_STEP=defaults['LOOKUP_STEP'], 
+    TEST_SIZE=defaults['TEST_SIZE'], 
+    N_LAYERS=defaults['N_LAYERS'], 
+    CELL=defaults['CELL'], 
+    UNITS=defaults['UNITS'], 
+    DROPOUT=defaults['DROPOUT'], 
+    BIDIRECTIONAL=defaults['BIDIRECTIONAL'], 
+    LOSS=defaults['LOSS'],
+    OPTIMIZER=defaults['OPTIMIZER'], 
+    BATCH_SIZE=defaults['BATCH_SIZE'], 
+    EPOCHS=defaults['EPOCHS'],
+    PATIENCE=defaults['PATIENCE'],
+    SAVELOAD=defaults['SAVELOAD']):
+
+    data, model, test_acc, valid_acc, train_acc, mae = make_neural_net(
+        ticker, end_date=end_date, 
+        N_STEPS=N_STEPS, LOOKUP_STEP=LOOKUP_STEP, TEST_SIZE=TEST_SIZE, 
+        N_LAYERS=N_LAYERS, CELL=CELL, UNITS=UNITS, DROPOUT=DROPOUT, 
+        BIDIRECTIONAL=BIDIRECTIONAL, LOSS=LOSS, OPTIMIZER=OPTIMIZER, 
+        BATCH_SIZE=BATCH_SIZE, EPOCHS=EPOCHS, PATIENCE=PATIENCE,
+        SAVELOAD=SAVELOAD
+    )
+
 
 def make_neural_net(ticker, end_date=None, 
     N_STEPS=defaults['N_STEPS'], 
@@ -92,8 +122,9 @@ def make_neural_net(ticker, end_date=None,
     OPTIMIZER=defaults['OPTIMIZER'], 
     BATCH_SIZE=defaults['BATCH_SIZE'], 
     EPOCHS=defaults['EPOCHS'],
-    PATIENCE=defaults['PATIENCE']):
-#description of these parameters located inside environment.py
+    PATIENCE=defaults['PATIENCE'],
+    SAVELOAD=defaults['SAVELOAD']):
+    #description of these parameters located inside environment.py
 
     tf.keras.backend.clear_session()
     tf.config.optimizer.set_jit(True)
@@ -113,7 +144,7 @@ def make_neural_net(ticker, end_date=None,
     model_name = f"{date_now}_{ticker}-{LOSS}-{OPTIMIZER}-{CELL.__name__}-seq-{N_STEPS}-step-{LOOKUP_STEP}-layers-{N_LAYERS}-units-{UNITS}"
     if BIDIRECTIONAL:
         model_name += "-b"
-    # create these folders if they does not exist
+    # create these folders if they do not exist
     results_folder = 'results'
     if not os.path.isdir(results_folder):
        os.mkdir(results_folder)
@@ -126,7 +157,10 @@ def make_neural_net(ticker, end_date=None,
 
     logs = "logs/" + get_time_string()
 
-    checkpointer = ModelCheckpoint(os.path.join("results", model_name + ".h5"), save_weights_only=True, save_best_only=True, verbose=1)
+    if SAVELOAD:
+        checkpointer = ModelCheckpoint(model_saveload_directory + '/' + ticker + ".h5", save_weights_only=False, save_best_only=True, verbose=1)
+    else:    
+        checkpointer = ModelCheckpoint(os.path.join("results", model_name + ".h5"), save_weights_only=True, save_best_only=True, verbose=1)
     
     if save_logs:
         tboard_callback = tf.keras.callbacks.TensorBoard(log_dir=logs, profile_batch='100,200') 
@@ -147,26 +181,29 @@ def make_neural_net(ticker, end_date=None,
 
     
     #before testing, no shuffle
-    data, train, valid, test = load_data(
-        ticker, N_STEPS, lookup_step=LOOKUP_STEP, 
-        test_size=TEST_SIZE, shuffle=False, batch_size=BATCH_SIZE,
-        end_date=end_date
-    )
+    if SAVELOAD:
+        test_acc, valid_acc, train_acc, mae = 0    
+    else:    
+        data, train, valid, test = load_data(
+            ticker, N_STEPS, lookup_step=LOOKUP_STEP, 
+            test_size=TEST_SIZE, shuffle=False, batch_size=BATCH_SIZE,
+            end_date=end_date
+        )
 
-    model_path = os.path.join("results", model_name + ".h5")
-    model.load_weights(model_path)
+        model_path = os.path.join("results", model_name + ".h5")
+        model.load_weights(model_path)
 
-    mse, mae = model.evaluate(test, verbose=0)
-    mae = data["column_scaler"][test_var].inverse_transform([[mae]])[0][0]
-    
-    delete_files_in_folder(results_folder)
-    os.rmdir(results_folder)
-    
-    y_valid_real, y_valid_pred = return_real_predict(model, data["X_valid"], data["y_valid"], data["column_scaler"][test_var])
-    valid_acc = get_accuracy(y_valid_real, y_valid_pred, LOOKUP_STEP)
-    y_test_real, y_test_pred = return_real_predict(model, data["X_test"], data["y_test"], data["column_scaler"][test_var])
-    test_acc = get_accuracy(y_test_real, y_test_pred, LOOKUP_STEP)
-    y_train_real, y_train_pred = return_real_predict(model, data["X_train"], data["y_train"], data["column_scaler"][test_var])
-    train_acc = get_accuracy(y_train_real, y_train_pred, LOOKUP_STEP)
+        mse, mae = model.evaluate(test, verbose=0)
+        mae = data["column_scaler"][test_var].inverse_transform([[mae]])[0][0]
+        
+        delete_files_in_folder(results_folder)
+        os.rmdir(results_folder)
+        
+        y_valid_real, y_valid_pred = return_real_predict(model, data["X_valid"], data["y_valid"], data["column_scaler"][test_var])
+        valid_acc = get_accuracy(y_valid_real, y_valid_pred, LOOKUP_STEP)
+        y_test_real, y_test_pred = return_real_predict(model, data["X_test"], data["y_test"], data["column_scaler"][test_var])
+        test_acc = get_accuracy(y_test_real, y_test_pred, LOOKUP_STEP)
+        y_train_real, y_train_pred = return_real_predict(model, data["X_train"], data["y_train"], data["column_scaler"][test_var])
+        train_acc = get_accuracy(y_train_real, y_train_pred, LOOKUP_STEP)
 
     return data, model, test_acc, valid_acc, train_acc, mae
